separate and update admin routes, update product service and store

+++ backend/src/services/product.service.ts
@@ -1,28 +1,17 @@
 import { ProductRepository } from '../repositories/product.repository.js';
-import { CreateProductDto, GetProductsQuery } from '../validators/product.validator.js';
+import { CreateProductDto, GetProductsQuery, UpdateProductDto } from '../validators/product.validator.js';
 import { createUniqueSlug } from '../utils/slugify.js';
-import { NotFoundError } from '../utils/errors.js';
-import { ReqUser } from '../types/token.types.js';
-import { Role } from '@prisma/client';
+import { NotFoundError, ValidationError } from '../utils/errors.js';
 
 export class ProductService {
-	static getProducts = async (user: ReqUser | undefined, params: GetProductsQuery) => {
+	static getProductsPublic = async (params: GetProductsQuery) => {
 		const { page, limit } = params;
 		const skip = (page - 1) * limit;
 
-		let products;
-		let total;
-		if (user?.role === Role.ADMIN || user?.role === Role.SUPER_ADMIN) {
-			[products, total] = await Promise.all([
-				ProductRepository.findManyAdmin({ skip, take: limit }),
-				ProductRepository.count(),
-			]);
-		} else {
-			[products, total] = await Promise.all([
-				ProductRepository.findManyPublic({ skip, take: limit }),
-				ProductRepository.count(),
-			]);
-		}
+		const [products, total] = await Promise.all([
+			ProductRepository.findManyPublic({ skip, take: limit }),
+			ProductRepository.count(),
+		]);
 
 		return {
 			products,
@@ -35,13 +24,38 @@ export class ProductService {
 		};
 	};
 
-	static getProductById = async (user: ReqUser | undefined, id: number) => {
-		let product;
-		if (user?.role === Role.ADMIN || user?.role === Role.SUPER_ADMIN) {
-			product = await ProductRepository.findProductByIdAdmin(id);
-		} else {
-			product = await ProductRepository.findProductByIdPublic(id);
+	static getProductByIdPublic = async (id: number) => {
+		const product = await ProductRepository.findProductByIdPublic(id);
+
+		if (!product) {
+			throw new NotFoundError(`User with id ${id} not found`);
 		}
+
+		return product;
+	};
+
+	static getProductsAdmin = async (params: GetProductsQuery) => {
+		const { page, limit } = params;
+		const skip = (page - 1) * limit;
+
+		const [products, total] = await Promise.all([
+			ProductRepository.findManyAdmin({ skip, take: limit }),
+			ProductRepository.count(),
+		]);
+
+		return {
+			products,
+			pagination: {
+				page: page,
+				limit: limit,
+				total: total,
+				totalPages: Math.ceil(total / limit),
+			},
+		};
+	};
+
+	static getProductByIdAdmin = async (id: number) => {
+		const product = await ProductRepository.findProductByIdAdmin(id);
 
 		if (!product) {
 			throw new NotFoundError(`User with id ${id} not found`);
@@ -57,8 +71,27 @@ export class ProductService {
 		});
 	};
 
-	static deleteProduct = async (user: ReqUser | undefined, id: number) => {
-		await this.getProductById(user, id);
+	static updateProduct = async (productData: UpdateProductDto, id: number) => {
+		if (!productData) {
+			throw new ValidationError('No product data');
+		}
+
+		const product = await this.getProductByIdAdmin(id);
+
+		if (!product) {
+			throw new NotFoundError(`User with id ${id} not found`);
+		}
+
+		return ProductRepository.updateProduct(productData, id);
+	};
+
+	static deleteProduct = async (id: number) => {
+		const product = await this.getProductByIdAdmin(id);
+
+		if (!product) {
+			throw new NotFoundError(`User with id ${id} not found`);
+		}
+
 		await ProductRepository.deleteProduct(id);
 	};
 }
